+++ CHANGES.rst
@@ -5,6 +5,14 @@ Changelog
 0.4dev
 ------
 
+- Change browser view and adapter regitrations from IPloneSiteRoot to
+  `zope.component.interfaces.ISite`. That's needed for Lineage compatibility.
+  [thet]
+
+- Integrate @@showorder view to access information for a specific order for
+  anonymous users by giving the ordernumber and email as credentials.
+  [thet]
+
 - Fix mail sending for AT based buyable items.
   [rnix]